ragflow/requirements.txt
neumai==0.0.38
google-cloud-aiplatform==1.42.1
pandas==2.1.0
google-cloud-storage==2.14.0
unstructured[all-docs]==0.12.2
unstructured[local-inference]==0.12.2
poppler-utils==0.1.0
psycopg2==2.9.9
sqlalchemy==2.0.27
fitz==0.0.1.dev2
celery==5.3.6
redis==5.0.1
gevent==24.2.1
gcsfs==2023.6.0
flask==3.0.2
PyPDF2==3.0.1
flower==2.0.1
google-cloud-monitoring
streamlit==1.26.0
langchain==0.0.335
trubrics==1.4.3
streamlit-feedback==0.1.3
requests==2.31.0
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
spacy==3.7.4
datasets==2.17.0
span-marker==1.5.0
scikit-learn==1.2.2
sqlalchemy[asyncio]
pgvector==0.2.5
psycopg2-binary==2.9.9
asyncpg==0.29.0
wandb==0.16.2
ragas==0.1.1
nest-asyncio==1.6.0
llama_index==0.10.6
llama-index-embeddings-Gemini==0.1.1
llama-index-vector-stores-postgres==0.1.1
llama-index-callbacks-wandb==0.1.1
llama-index-readers-file==0.1.3
llama-index-llms-Vertex==0.1.1
llama-index-llms-ollama==0.1.2
ollama==0.1.6
llama-index-embeddings-huggingface
